requirements.txt
Django==2.0.2
django-crispy-forms
django-import-export
django-reversion
django-formtools==2.1
django-registration
django-rest-swagger
djangorestframework
future==0.15.2
httplib2==0.9.2
pymysql
requests
reversion
django-cors-headers
django-cors-middleware
qiniu
cycler== 0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2+mkl
pyparsing == 2.4.0
python-dateutil== 2.8.0
scikit-learn == 0.20.3
scipy ==1.2.1
six== 1.12.0
pandas
apscheduler == 2.1.2